requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2.4
django-heroku==0.3.1
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.12.4
dnspython==2.1.0
gunicorn==20.1.0
mongoengine==0.23.1
psycopg2==2.9.1
pymongo==3.11.4
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
